fact_verifier/requirements.txt
# Core NLP and models
spacy==3.7.2
en-core-web-sm@https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
sentence-transformers==2.7.0
transformers==4.45.2
torch>=2.2.0

# Retrieval and parsing
wikipedia==1.4.0
readability-lxml==0.8.1
beautifulsoup4==4.12.3
lxml==5.3.0
lxml_html_clean==0.2.0
httpx==0.27.2
requests==2.32.3

# Utils
numpy==1.26.4
scipy==1.13.1
scikit-learn==1.5.2
pydantic==2.9.2

# Apps
fastapi==0.115.2
uvicorn[standard]==0.31.1
streamlit==1.39.0

# Testing
pytest==8.2.1
pytest-asyncio==0.24.0

